Move production types to economy/production/
