+++ src/openvic-simulation/economy/BuildingType.hpp
@@ -1,11 +1,12 @@
 #pragma once
 
-#include "openvic-simulation/misc/Modifier.hpp"
 #include "openvic-simulation/economy/GoodDefinition.hpp"
-#include "openvic-simulation/economy/ProductionType.hpp"
+#include "openvic-simulation/economy/production/ProductionType.hpp"
+#include "openvic-simulation/misc/Modifier.hpp"
 #include "openvic-simulation/types/Date.hpp"
 #include "openvic-simulation/types/IdentifierRegistry.hpp"
 #include "openvic-simulation/types/fixed_point/FixedPoint.hpp"
+
 
 namespace OpenVic {
