+++ src/openvic-simulation/economy/EconomyManager.hpp
@@ -2,8 +2,10 @@
 
 #include "openvic-simulation/economy/BuildingType.hpp"
 #include "openvic-simulation/economy/GoodDefinition.hpp"
-#include "openvic-simulation/economy/ProductionType.hpp"
+#include "openvic-simulation/economy/production/ProductionType.hpp"
+
 #include "openvic-dataloader/v2script/Parser.hpp"
+
 
 namespace OpenVic {
 	struct EconomyManager {
