+++ src/openvic-simulation/economy/production/ArtisanalProducer.cpp
