+++ src/openvic-simulation/economy/production/ArtisanalProducer.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "openvic-simulation/economy/GoodDefinition.hpp"
-#include "openvic-simulation/economy/ProductionType.hpp"
+#include "openvic-simulation/economy/production/ProductionType.hpp"
 #include "openvic-simulation/types/fixed_point/FixedPoint.hpp"
 #include "openvic-simulation/utility/Getters.hpp"
 
@@ -15,10 +15,8 @@ namespace OpenVic {
 
 	public:
 		ArtisanalProducer(
-			ProductionType const& new_production_type,
-			GoodDefinition::good_definition_map_t&& new_stockpile,
-			fixed_point_t new_current_production,
-			GoodDefinition::good_definition_map_t&& new_current_needs
+			ProductionType const& new_production_type, GoodDefinition::good_definition_map_t&& new_stockpile,
+			fixed_point_t new_current_production, GoodDefinition::good_definition_map_t&& new_current_needs
 		);
 	};
 }
