+++ src/openvic-simulation/economy/production/ResourceGatheringOperation.cpp
